Use WindowKit for presentation
